+++ src/index.tsx
@@ -13,23 +13,26 @@ import "./styles/global-style.css";
 import "react-toastify/dist/ReactToastify.css";
 import strings from "./localization";
 import LoadingSpinner from "./modules/components/loading-spinner";
+import ErrorBoundary from "./modules/components/error-boundaries";
 
 let persistor = persistStore(store);
 
 ReactDOM.render(
   <React.StrictMode>
-    <Provider store={store}>
-      <PersistGate loading={<LoadingSpinner />} persistor={persistor}>
-        <ThemeProvider theme={theme}>
-          <LocalizationProvider
-            dateAdapter={DateAdapter}
-            locale={strings.getInterfaceLanguage()}
-          >
-            <App className="App" />
-          </LocalizationProvider>
-        </ThemeProvider>
-      </PersistGate>
-    </Provider>
+    <ErrorBoundary>
+      <Provider store={store}>
+        <PersistGate loading={<LoadingSpinner />} persistor={persistor}>
+          <ThemeProvider theme={theme}>
+            <LocalizationProvider
+              dateAdapter={DateAdapter}
+              locale={strings.getInterfaceLanguage()}
+            >
+              <App className="App" />
+            </LocalizationProvider>
+          </ThemeProvider>
+        </PersistGate>
+      </Provider>
+    </ErrorBoundary>
   </React.StrictMode>,
   document.getElementById("root")
 );