+++ src/modules/components/footer/style.tsx
@@ -5,7 +5,6 @@ export const FooterRoot = styled((props: any) => <div {...props} />)(
   (_: any) => `
     background-color: white;
     width: 100%;
-    margin-top: 150px;
     border-top: 1px solid #e9e9e9;
     height: ${FOOTER_HEIGHT};
   `